Refactor constants modules to a higher level to avoid circular imports
