+++ src/distilabel/pipeline/step_wrapper.py
@@ -16,9 +16,9 @@ import traceback
 from queue import Queue
 from typing import Any, Dict, List, Optional, Union, cast
 
+from distilabel.constants import LAST_BATCH_SENT_FLAG
 from distilabel.llms.mixins.cuda_device_placement import CudaDevicePlacementMixin
 from distilabel.pipeline.batch import _Batch
-from distilabel.pipeline.constants import LAST_BATCH_SENT_FLAG
 from distilabel.pipeline.typing import StepLoadStatus
 from distilabel.steps.base import GeneratorStep, Step, _Step
